+++ lib/upgrader.dart
@@ -1,11 +1,14 @@
 /*
- * Copyright (c) 2018 Larry Aasen. All rights reserved.
+ * Copyright (c) 2018-2021 Larry Aasen. All rights reserved.
  */
 
 library upgrader;
 
+export 'src/alert_style_widget.dart';
 export 'src/appcast.dart';
 export 'src/itunes_search_api.dart';
 export 'src/upgrade_alert.dart';
+export 'src/upgrade_base.dart';
+export 'src/upgrade_card.dart';
 export 'src/upgrade_messages.dart';
 export 'src/upgrader.dart';